+++ spec/support/schema.rb
@@ -1,9 +1,30 @@
 require 'active_record'
 
-ActiveRecord::Base.establish_connection(
-  adapter: 'sqlite3',
-  database: ':memory:'
-)
+if ENV['ACTIVE_RECORD_ADAPTER'] == 'mysql'
+  puts 'Running on MySQL...'
+  ActiveRecord::Base.establish_connection(
+    adapter: 'mysql2',
+    host: ENV['DB_HOST'] || '127.0.0.1',
+    username: ENV['DB_USERNAME'] || 'root',
+    password: ENV['DB_PASSWORD'],
+    database: 'acts_as_nosql'
+  )
+elsif ENV['ACTIVE_RECORD_ADAPTER'] == 'postgresql'
+  puts 'Running on PostgreSQL...'
+  ActiveRecord::Base.establish_connection(
+    adapter: 'postgresql',
+    database: 'acts_as_nosql',
+    host: ENV['DB_HOST'] || '127.0.0.1',
+    username: ENV['DB_USERNAME'] || ENV['POSTGRES_USER'] || 'postgres',
+    password: ENV['DB_PASSWORD'] || ENV['POSTGRES_PASSWORD']
+  )
+else
+  puts 'Running on SQLite...'
+  ActiveRecord::Base.establish_connection(
+    adapter: 'sqlite3',
+    database: ':memory:'
+  )
+end
 
 class Common < ActiveRecord::Base
   self.abstract_class = true